feat: add method in class for adding brands
