+++ online-store/src/styles/blocks/category.scss
@@ -4,6 +4,7 @@
 	width: 52px;
 	border: 1px solid gray;
 	padding: 2px;
+	margin-bottom: 2px;
 
 	&__label {
 		display: flex;
