+++ online-store/src/styles/index.scss
@@ -16,3 +16,17 @@ body {
 @import 'blocks/category';
 @import 'blocks/label';
 
+.brand {
+	display: flex;
+	flex-direction: column;
+	width: 52px;
+	border: 1px solid gray;
+	padding: 2px;
+	margin-bottom: 2px;
+
+	&__label {
+		display: flex;
+		justify-content: space-between;
+	}
+}
+
